I've formulated a plan at least.
I think because I'm not using this as a blog with dated posts, the .md files aren't getting rendered to html on the GitHub servers.

My plan for now is to repurpose this gh-pages branch to contain *only* the _site directory and keep raw development materials on the the master branch.

It's going to take a rebase and few commits to get there.  Let's get started!