give each section a description
This commit gives a "description" attribute to most chapters' sections and displays that description on the index page.  This should help people understand what the chapter will be about before they click on it.

This commit also adds some general documentation updates to Sprite and Quad.

Finally, this commit updates the gemfile.