misc cleanup
This commit breaks CommandChaining out into its own chapter.  It was previously a subsection of the chapter on Quads.  This new chapter, along with the chapter on Debugging now comprise a new bundle of chapters on Best Practices.

This commit also cleans up/adds to the explanations in Foreward and Quad.  Finally, the gemfile has been updated.