add TopBar navigation
This commit adds the external resources and code necessary for a Zurb Foundation 6 topbar in the main layout.  Yay!

It's a little bit broken right now in mobile layout; I'll attempt to fix that in a later commit.

I also need to fill the new "Other Resources" page with content!