transition from Jekyll to React
This commit completely overhauls the site to be a React/React Router SPA rather than a Jekyll MPA.

I learned a lot about ES6, npm, and React in the last 48 hours.  There's still a lot to learn, but this change feels good, overall.

Now to push to GitHub pages and see what breaks!