scroll window after constructor, if needed
If the url string has a hash, the page should automatically scroll to the appropriate element's y-offset value.  Normally, the browser handles if element id's are set up (they are, here), but the double AJAX call and XML parsing at page load means that the document takes a while to be populated with content to scroll to.

This commit adds a few lines of JavaScript that will check for the presence of a hash in the url string and scroll to the appropriate element AFTER render() is done.