less hackish window scroll onhashchange
This seems to produce more consistent y-offsets and is wrapped in a function, scroll_window_after_hashchange, so it can be called after the document is ready, as well as after hashchange.

This also (hopefully) improves the CSS that makes <br> tags within method-description <span> elements behave more like <p> elements in a seemingly cross-browser compatible way.