hide table header for Constants if it has 0 rows
After filtering the page, it's possible for the Constants table to have no rows of data.

Previously, the table's <thead> element would always display, even if the table contained 0 rows of data.  This was potentially confusing, so this commit changes that behavior; the entire table is hidden if it currently has 0 rows of data.