move text filter input to top of layout in mobile
The text filter <input> element is normally part of the sidebar, but the sidebar gets pushed to the bottom of the page for mobile due to lack of horizontal space.  This made getting to the text filter input extremely inconvenient on mobile!

This commit adds a second LuaAPIFilter component, now directly into the LuaAPI component, and uses bootstrap syntax to show it for mobile and hide it for "md" and larger.  Part of this involved refactoring the timeout code from App.js into LuaAPIFilter.jsx where it probably belonged all along.

Hopefully this is helpful to anyone attempting to use the Lua API on mobile.