hopefully fix the site for Edge
I tested the site in MS Edge today and discovered that any page that tried to access the classList property of a DOM element crashed and didn't load at all.  (I think that was the problem, anyway.  I didn't have time to properly investigate/debug.)

Thus, this commit removes all uses of classList and replaces them with corresponding jQuery calls.  I recently added jQuery as a library while working on the new LuaAPI, so I was now able to use it elsewhere.

Hopefully the site works in Edge now!