style refresh
Side navigation moved to the left to mimic Create React App documentation.

Additionally, side navigation now hides when the window is smaller than bootstrap's medium, and a secondary mneu-toggle-button becomes visible, fixed to the bottom-right of the screen.

The hope is that this secondary button makes it easier to access nav on mobile when scrolling around in very long pages (e.g., the API doc).