npm update to jquery and bootstrap
jQuery recently released v3.5.0 with a security patch for the internal
handling of closing tags.  See: https://blog.jquery.com/2020/04/10/jquery-3-5-0-released/

Allowing npm to upgrade jQuery to 3.5.0 revealed a bug with the current
version of bootstrap that would cause this SPA to crash when a user
clicked to collapse something, e.g. a sidebar section.

Bootstrap addressed this in https://github.com/twbs/bootstrap/pull/30559
but there is not yet a Bootstrap release that includes this.  I'd previously
configured package.json to use release versions of Bootstrap (and all
dependencies).

To get the new Bootstrap fixes, I changed the package.json to use twbs's
GitHub repo for Bootstrap as outlined here:
https://docs.npmjs.com/files/package.json#github-urls

The update to Bootstrap has resulted in some minor oddities with this
app, but I'll address those in a separate commit.