ensure "Lua for SM5" in header is always white
Because "Lua for SM5" (and the SM5 arrow logo) are wrapped in a
Bootstrap nav-link component, the text appears gray when the URL doesn't
match "/" and the anchor is no longer considered active.

This is a quick hack to force the "Lua for SM5" text to always appear as
an active anchor with white text.  A better solution is probably to not
use a nav-link component for it.