make LuaAPIFilter sticky to top in mobile
This should make it possible to use the LuaAPI's textfield filter when
scrolling around on mobile device by making it sticky to top.

The implementation here is a bit of hack (three forced line breaks so
that the textfield doesn't get stuck "under" the navbar and is still
visible) but is better than before.  I can work on a better fix once the
beta warning box is safe to remove (someday).