check for <Link> elements in class <Description>s
Previously, I wasn't checking for the presence of <Link> elements within
<Description> elements that are direct children of Actor classes.

The <Link>s wouldn't be converted to anchors and would remain in the DOM
as-is.  Firefox, Chrome, and Safari opted to not render them at all,
resulting in users seeing sentences like:

These commands require a RageFile handle. You can create one using .

This should be handled now.

I still need to fix the case where multiple <Link> elements are matched
in a single <Description> elements, but this is still better than it
was.