convert <Link> to <a> following Lua.xsl specs
This attempts to follow along with the rules specified in Lua.xsl for
converting <Link> elements with web-browser compliant <a> elements.
Previously, I (hadn't yet found these rules and) had been doing my own
thing based on observing what Lua.xml looked like in a web browser.

I chucked all the RegExps I was using to parse XML (:nervous laughter:)
and settled on using jQuery for now.  It's already a dependency and will
be for a while longer, so I might as well use it to do what it does
best.