get methods from Lua.xml, not LuaDocumentation.xml
We have two xml files to retrieve data from, Lua.xml and
LuaDocumentation.xml.

Most Actor classes are *described* in LuaDocumentation.xml, but not all.

Lua.xml, on the other hand, is supposed to be autogenerated from the SM
engine using the command-line argument "--ExportLuaInformation"

All available classes and methods should be in Lua.xml, though not
everything has a corresponding explanation in LuaDocumentation.xml.

This commit changes the React app LuaAPI to show generate the list of
class methods from Lua.xml.