make Actors, Namespaces in API sidebar expandable
This works for now.  The React-specific code I've written in
LuaAPISidebar.jsx feels quite hacky, and suggests I need to learn more
about state and rendering in React apps.  But this works for now.

Now that Actor Classes and Namespaces in the sidebar are expandable,
there's not an obvious way to generate a URL that links to
/LuaAPI#Actors or /LuaAPI#Namespaces, particularly with how the mobile
navigation is configure to autohide on hash change...

I'll probably add clickable Octicons later.