more bootstrap fixes for sidebar
You can now click on an expandable sidebar section like "Classes" or
"Singletons" or "Enums" and it will both expand the section in the
sidebar and change the hash in the url, causing the main page to jump to
that section.

This should have been possible before with the way I had the anchors set
up, but I think bootstrap's collapse component may have been effective
preventing the hashchange event.