link to SM's GitHub for themeside GlobalFunctions
This is still largely proof-of-concept and not streamlined, but it's a
start.

GlobalFunctions.lua run in an SM5 theme will generate a simple csv file
of global functions using Lua's debug.getinfo utility.

The csv-parse library brought in via npm in the previous commit parses
that CSV and we can now link to GitHub for GlobalFunctions that are
defined in-theme.

Anything defined in the engine, GlobalFunction or otherwise, isn't
handled yet.  I'll need to look into something similar for C++ files,
preferably something Node-based so generating this data can be
integrated as an "npm run" scipt.

Still, this is a nice start.