sort classes, singletons, namespaces, enums
I was trying to find RageFileManager under the Singletons expandable
sidebar list, and noticed the list was not alphabetically sorted.  It
was using the sequence from Lua.xml.

This sorts classes, singletons, namespaces, and enums after retrieving
them from Lua.xml.

Global Functions and Constants are left as-is for now.  Neither has an
expandable list in the sidebar, and both need more careful consideration
than "alphabetize them just because."