sticky headers in LuaAPI
Problem:
If I control-F and search for "join", I'll eventually come across the
method for EnoughCreditsToJoin() which is under GAMESTATE.

GAMESTATE has so many methods, however, that if I suddenly jump to the
documentation for EnoughCreditsToJoin, there'd be no context for me know
it was a member function of the GAMESTATE singleton, requiring me to
scoll up until I found the section header.

Fix:
This commit applies CSS for position:sticky to API sections, so that
both methods (for Actor and GAMESTATE and NoteField and etc.) and Global
functions have some context no matter how you come across them.