remove react-router-dom as a dep
As https://github.com/quietly-turning/Lua-For-SM5/pull/22 pointed out,
react-router-dom is not necessary to keep in this project.  It was
intended as a temporary upgrade shim to ease the transition from router
v5 to router v6, ...which was a while ago.

So, let's remove it.  react-router is all that's needed.